chore: fix clippy lints
- registry: rank_search_results / list_tags use sort_by_key(Reverse(..)) instead
  of sort_by with a recomputed key (clearer; relevance score computed once per
  element rather than twice per comparison).
- resolver: use u64::div_ceil instead of the manual (n + d - 1) / d.
- hardware: drop the redundant `as i32` cast on libc::HOST_VM_INFO64.
- display: drop redundant .to_string() in a println! arg.

No behavior change; `cargo clippy --all-targets` is now clean.

+++ src/hardware.rs
@@ -415,7 +415,7 @@ fn darwin_free_ram_bytes() -> Option<u64> {
     let ret = unsafe {
         libc::host_statistics64(
             host,
-            libc::HOST_VM_INFO64 as i32,
+            libc::HOST_VM_INFO64,
             &mut vm_stat as *mut libc::vm_statistics64_data_t as *mut libc::integer_t,
             &mut count,
         )
